% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-retriever.R
\name{addin_run_retriever}
\alias{addin_run_retriever}
\title{Run Chat GPT with Retrieval}
\usage{
addin_run_retriever()
}
\value{
This function has no return value.
}
\description{
Run the ChatGPT shiny app with semantic search and document retrieval
}
\examples{
# Call the function as an RStudio addin
\dontrun{
addin_run_retriever()
}
}
