% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_scrape_pkgs.R
\name{addin_run_scrape_pkgs}
\alias{addin_run_scrape_pkgs}
\title{Addin to scrape installed packages}
\usage{
addin_run_scrape_pkgs()
}
\value{
No return value, called for side effects only.
}
\description{
Invokes RStudio addin functionality to scrape select installed packages and
create indices for use in the "Chat with Retrieval" application.
}
\note{
This addin requires RStudio to be available and will stop with an
error message if RStudio API is not accessible.
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# This function is typically run within RStudio as an Addin.
# It would not be called directly in the code.
addin_scrape_pkgs()
\dontshow{\}) # examplesIf}
}
