% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_select_pkgs.R
\name{addin_run_select_pkgs}
\alias{addin_run_select_pkgs}
\title{Run Package Selector App}
\usage{
addin_run_select_pkgs()
}
\value{
This function has no return value.
}
\description{
Run the package selector shiny app
}
