% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin-run-retriever.R
\name{chat_with_retrieval}
\alias{chat_with_retrieval}
\title{Run Chat with Retrieval}
\usage{
chat_with_retrieval()
}
\value{
This function has no return value.
}
\description{
Run the Chat with Retrieval shiny app
}
\examples{
# Call the function as an RStudio addin
\dontrun{
chat_with_retrieval()
}
}
