% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_code_addin.R
\name{comment_code_addin}
\alias{comment_code_addin}
\title{active voice Addin}
\usage{
comment_code_addin()
}
\description{
Call this function as a Rstudio addin to ask GPT to add comments to your code
}
