% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe.R
\name{create_index_from_audio}
\alias{create_index_from_audio}
\title{Create index from audio}
\usage{
create_index_from_audio(file_path, source, link = NULL, overwrite = FALSE)
}
\arguments{
\item{file_path}{Character string specifying the path to the audio file.}

\item{source}{Character string specifying the source of the audio.}

\item{link}{Character string specifying the link to the audio file
(optional).}

\item{overwrite}{Logical, whether to overwrite the existing index (default:
FALSE).}
}
\value{
The function writes an index in Parquet format to disk.
}
\description{
This function creates an index by first transcribing the audio file and then
creating the index.
}
