% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{create_index_from_pdf}
\alias{create_index_from_pdf}
\title{Create index from PDF}
\usage{
create_index_from_pdf(file_path, source, link = NULL, overwrite = FALSE)
}
\arguments{
\item{file_path}{Character string specifying the path to the PDF file.}

\item{source}{Character string specifying the source of the PDF.}

\item{link}{Character string specifying the link to the PDF file (optional).}

\item{overwrite}{Logical, whether to overwrite the existing index (default:
FALSE).}
}
\value{
The function writes an index in Parquet format to disk.
}
\description{
This function creates an index by first ingesting the PDF file and then
creating the index.
}
