% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_docs.R
\name{create_llama_index}
\alias{create_llama_index}
\title{Create Llama Index}
\usage{
create_llama_index(domain, index_type = "simple")
}
\arguments{
\item{domain}{character string specifying the domain to create an index for}

\item{index_type}{character string specifying the type of index to create}
}
\value{
the index created
}
\description{
Create a Llama index for a given domain
}
\examples{
\dontrun{
create_llama_index("movies")
create_llama_index("books", index_type = "faiss")
}
}
