% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe.R
\name{create_transcript}
\alias{create_transcript}
\title{Get Transcription from Audio}
\usage{
create_transcript(file_path, prompt = NULL, chunk_size = 120)
}
\arguments{
\item{file_path}{Character string specifying the path to the audio file.}

\item{prompt}{Character string specifying the prompt for transcription
(optional).}

\item{chunk_size}{Audio size in seconds.}
}
\value{
A character string containing the transcription text.
}
\description{
This function transcribes audio files and returns the transcription text.
}
