% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{delete_index}
\alias{delete_index}
\title{Delete an Index File}
\usage{
delete_index()
}
\description{
Interactively deletes a specified index file from a user-defined directory.
Presents the user with a list of available index files and prompts for confirmation
before deletion.
}
