% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-code.R
\name{extract_code_chunks}
\alias{extract_code_chunks}
\title{Extract Code Chunks from Text}
\usage{
extract_code_chunks(text)
}
\arguments{
\item{text}{A character vector containing the text from which code chunks
need to be extracted.}
}
\value{
A list containing the extracted code chunks.
}
\description{
This function extracts code chunks that are enclosed in triple backticks from
a given text input. It supports identifying code blocks that are indicated by
triple backticks followed by an optional language identifier. The function
returns the extracted code chunks.
}
\examples{
sample_text <- "Here is some text.
```
This is a code chunk.
```"
extract_code_chunks(sample_text)

}
