% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{ghost_writer}
\alias{ghost_writer}
\title{Writing Assistant}
\usage{
ghost_writer(
  service = getOption("gpttools.service", "openai"),
  stream = TRUE,
  where = "source"
)
}
\arguments{
\item{service}{The AI service to be used for the request. If not explicitly
provided, this defaults to the value set in
\code{getOption("gptstudio.service")}. If the option is not set, make sure to
provide this parameter to avoid errors.}

\item{stream}{A logical value indicating whether the interaction should be
treated as a stream for continuous interactions. If not explicitly
provided, this defaults to the value set in
\code{getOption("gptstudio.stream")}.}

\item{where}{A character string indicating the location or environment where
the chat is taking place. Options are \verb{c("console", "source", and "shiny")}. The
default is \code{""}, which means the chat is taking place in the R console.}
}
\description{
Writing Assistant
}
