% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_sitrep.R
\name{gpt_sitrep}
\alias{gpt_sitrep}
\title{GPT Sitrep}
\usage{
gpt_sitrep()
}
\value{
A message is returned to the user informing them of the status of the
OpenAI API, RStudio API, and gpttools settings.
}
\description{
Check the status of the OpenAI API and RStudio API, and inform
the user of the settings for gpttools.
}
\examples{
\dontrun{
gpt_sitrep()
}
}
