% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpttools-package.R
\docType{package}
\name{gpttools-package}
\alias{gpttools}
\alias{gpttools-package}
\title{gpttools: Extensions and Tools for gptstudio}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

gpttools is an R package that provides extensions to gptstudio to provide devtools-like functionality using the latest natural language processing (NLP) models. It is designed to make package development easier by providing a range of tools and functions that can be used to improve the quality of your package's documentation, testing, and maybe even functionality.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JamesHWade/gpttools}
  \item Report bugs at \url{https://github.com/JamesHWade/gpttools/issues}
}

}
\author{
\strong{Maintainer}: James Wade \email{github@jameshwade.com} (\href{https://orcid.org/0000-0002-9740-1905}{ORCID})

}
\keyword{internal}
