% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{ingest_pdf}
\alias{ingest_pdf}
\title{Ingest PDF and write to index}
\usage{
ingest_pdf(file_path, source, link = NULL)
}
\arguments{
\item{file_path}{Character string specifying the path to the PDF file.}

\item{source}{Character string specifying the source of the PDF.}

\item{link}{Character string specifying the link to the PDF file (optional).}
}
\value{
The function writes an index in Parquet format to disk.
}
\description{
This function reads the text from a PDF file, processes it, and writes the
result to an index.
}
