% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{load_index}
\alias{load_index}
\title{Load Index Data for a Domain}
\usage{
load_index(domain, local_embeddings = FALSE)
}
\arguments{
\item{domain}{A character string indicating the name of the domain.}

\item{local_embeddings}{A logical indicating whether to load the local
embeddings or the OpenAI embeddings. Defaults to FALSE.}
}
\value{
A data frame containing the index data for the specified domain.
}
\description{
This function loads the index data for a given domain from a parquet file.
}
\examples{
\dontrun{
load_index("example_domain")
}
}
