% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_docs.R
\name{load_llama_index}
\alias{load_llama_index}
\title{Load Llama Index}
\usage{
load_llama_index(domain)
}
\arguments{
\item{domain}{The domain to load the index from.}
}
\value{
A GPTSimpleVectorIndex object.
}
\description{
This function loads a Llama index from disk.
}
