% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_docs.R
\name{process_text_files}
\alias{process_text_files}
\title{Process text files}
\usage{
process_text_files(domain)
}
\arguments{
\item{domain}{Character. The name of the directory containing the text files.}
}
\value{
A data frame containing the processed text files
}
\description{
Process text files
}
