% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{query_index}
\alias{query_index}
\title{Query an Index}
\usage{
query_index(index, query, history, task = "Context Only", k = 4)
}
\arguments{
\item{index}{A pre-built index of text data.}

\item{query}{A character string representing the question or prompt to query
the index with.}

\item{history}{A list of the previous chat responses}

\item{task}{A character string indicating the task to perform, such as
"conservative q&a".}

\item{k}{An integer specifying the number of top matches to retrieve.}
}
\value{
A list containing the instructions for answering the question, the
context in which the question was asked, and the suggested answer.
}
\description{
This function queries an index with a given question or prompt and returns a
set of suggested answers.
}
\examples{
\dontrun{
index <- build_index(data)
query_index(index, "What is the capital of France?")
}
}
