% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_docs.R
\name{query_llama_index}
\alias{query_llama_index}
\title{Query the Llama Index}
\usage{
query_llama_index(index, query)
}
\arguments{
\item{index}{The Llama Index object.}

\item{query}{The query string.}
}
\value{
The result of the query.
}
\description{
This function queries the Llama Index using a query string.
}
