% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{read_history}
\alias{read_history}
\title{Read chat history from a file}
\usage{
read_history(file_name = "chat_history")
}
\arguments{
\item{file_name}{Name of the chat history file to read from. Default is
"chat_history".}
}
\value{
A dataframe containing the chat history or NULL if the file doesn't
exist.
}
\description{
Read chat history from a file
}
\examples{
\dontrun{
# Read chat history from the default file
history <- read_history()
}
}
