% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest-docs.R
\name{remove_lines_and_spaces}
\alias{remove_lines_and_spaces}
\title{Remove new lines from a character vector.}
\usage{
remove_lines_and_spaces(serie)
}
\arguments{
\item{serie}{A character vector.}
}
\value{
A character vector with all new line characters removed.
}
\description{
Removes all new line characters from a character vector.
}
