% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_docs.R
\name{remove_new_lines}
\alias{remove_new_lines}
\title{Remove new lines from a character vector.}
\usage{
remove_new_lines(serie)
}
\arguments{
\item{serie}{A character vector.}
}
\value{
A character vector with all new line characters removed.
}
\description{
Removes all new line characters from a character vector.
}
\examples{
remove_new_lines(c("Hello\nWorld", "Foo\nBar"))
}
