% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repair_index_names}
\alias{repair_index_names}
\title{Repair Index Names}
\usage{
repair_index_names(local = TRUE)
}
\arguments{
\item{local}{logical; if TRUE, the function operates on the local index
directory, otherwise it uses the main index directory.}
}
\value{
The function does not return a value, but it modifies the index files
in place by either adding a "name" column based on automated generation or
user input.
}
\description{
This function scans index files in a specified directory, checks if they lack
a "name" column, and attempts to repair them by adding a "name" column based
on the file name or user input.
}
\examples{
\dontrun{
repair_index_names(TRUE)
repair_index_names(FALSE)
}

}
