% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_document_data.R
\name{run_document_data}
\alias{run_document_data}
\title{Shiny app that supports Document Data addin}
\usage{
run_document_data()
}
\value{
Nothing is returned, a shiny app is run
}
\description{
Shiny app that supports Document Data addin
}
