% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-code.R
\name{run_extracted_code}
\alias{run_extracted_code}
\title{Execute and Present Extracted Code Chunks}
\usage{
run_extracted_code(code)
}
\arguments{
\item{code}{A character vector of code to be executed and presented. The code
is run with \code{reprex::reprex()}, capturing the outcome for presentation.}
}
\value{
An HTML-formatted string that includes the collapsible section for
the code's output. This HTML string can be embedded in web pages, markdown
documents, or viewed directly in browsers. The function is designed to work
seamlessly in interactive environments, hence it might return the output
invisibly.
}
\description{
This function takes extracted code chunks and utilizes the
\code{reprex} packages to run them. It then formats the output into a collapsible
HTML section for easy sharing and viewing. Images within the output are
converted to clickable links to enhance accessibility and compactness of the
presentation.
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
code <- "plot(1:10)"
html_output <- run_extracted_code(code)
# The returned value is an HTML string with the plotted output presented within a
# collapsible section and any images replaced by links.
\dontshow{\}) # examplesIf}
}
