% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_scrape_pkgs.R
\name{run_select_pkgs_app}
\alias{run_select_pkgs_app}
\title{Run a Shiny App to Select and Save Installed Packages}
\usage{
run_select_pkgs_app()
}
\value{
None The function is used for its side effect of launching a Shiny
app and doesn't return anything.
}
\description{
This function launches a Shiny application that allows users to select from a
list of installed packages and save their selections.
}
\details{
The application provides a sidebar for package selection and an
action button to save the selected packages. It displays the selected
packages in a data table.
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
run_select_pkgs_app()
\dontshow{\}) # examplesIf}
}
