% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{save_user_config}
\alias{save_user_config}
\title{Save user configuration settings for gpttools}
\usage{
save_user_config(
  service = "openai",
  model = "gpt-4-1106-preview",
  task = "Permissive Chat",
  embeddings = "local",
  k_context = 4,
  k_history = 4,
  save_history = TRUE,
  sources = "All",
  persist = TRUE
)
}
\arguments{
\item{service}{The name of the service to use, default is "openai".}

\item{model}{The model to use, default is "gpt-4-1106-preview".}

\item{task}{The task to perform, default is "Permissive Chat".}

\item{embeddings}{The location of embeddings, default is "local".}

\item{k_context}{The amount of context to keep, default is 4.}

\item{k_history}{The amount of history to keep, default is 4.}

\item{save_history}{Logical indicating whether history should be saved,
default is TRUE.}

\item{sources}{The sources to use, default is "All".}

\item{persist}{Logical indicating whether to persist the settings, default
is TRUE.}
}
\value{
Invisible NULL.
}
\description{
This function saves the user's configuration settings for gpttools service
to a YAML file.
}
