% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-index.R
\name{scrape_pkg_sites}
\alias{scrape_pkg_sites}
\title{Scrape packaging sites}
\usage{
scrape_pkg_sites(
  sites = get_pkgs_to_scrape(local = TRUE),
  service = "local",
  index_create = TRUE,
  overwrite = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{sites}{A data frame containing the package sites to be scraped. If not
provided, it defaults to \code{get_pkgs_to_scrape(local = TRUE)}.}

\item{service}{The service to be used for scraping, defaults to "local".}

\item{index_create}{Logical indicating whether to create an index, defaults
to \code{TRUE}.}

\item{overwrite}{Logical indicating whether to overwrite existing content,
defaults to \code{TRUE}.}

\item{parallel}{Logical indicating whether to use parallel processing,
defaults to \code{FALSE}.}
}
\value{
Invisible \code{NULL}. The function is called for its side effects.
}
\description{
Scrape packaging sites
}
\details{
This function scrapes the websites for the packages specified in the
\code{sites} dataframe. If \code{sites} is empty, it alerts the user with no packages
to scrape and returns \code{NULL} invisibly. If the user confirms to proceed, it
scrapes each package site using the supplied details.
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
scrape_pkg_sites()
\dontshow{\}) # examplesIf}
}
