% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest-docs.R
\name{scrape_url}
\alias{scrape_url}
\title{Scrape text from a URL}
\usage{
scrape_url(url)
}
\arguments{
\item{url}{A character string containing a valid URL.}
}
\value{
A character vector containing the text from the URL.
}
\description{
This function scrapes the text from a URL and returns a character vector.
}
