% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_to_function_addin.R
\name{script_to_function_addin}
\alias{script_to_function_addin}
\title{Convert Script to Function Addin}
\usage{
script_to_function_addin()
}
\description{
"convert this R code into an R function"
}
