% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_user_config}
\alias{set_user_config}
\title{Set user configuration settings for gpttools from a file}
\usage{
set_user_config(path = NULL)
}
\arguments{
\item{path}{The path to the config YAML file. If NULL, the default path is
used.}
}
\value{
Invisible NULL.
}
\description{
This function sets the user's configuration settings for gpttools by reading
a YAML file. If no path is provided, it will look for the file in the default
R user config directory.
}
