% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_better_code.R
\name{suggest_code_improvements}
\alias{suggest_code_improvements}
\title{Suggest code improvements}
\usage{
suggest_code_improvements()
}
\description{
Prompt: "I'm writing R code for use in a package. How can I improve this
code:"
}
