% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_suggest_test.R, R/suggest_test_addin.R
\name{suggest_unit_test_addin}
\alias{suggest_unit_test_addin}
\title{Suggest a unit test for a function}
\usage{
suggest_unit_test_addin()

suggest_unit_test_addin()
}
\description{
Prompt: "Suggest a unit test for this function using the testthat package"

Prompt: "Suggest a unit test for this function using the testthat package"
}
