% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_add_roxygen.R
\name{add_roxygen_addin}
\alias{add_roxygen_addin}
\title{Add Roxygen documentation to a function}
\usage{
add_roxygen_addin()
}
\value{
NULL (nothing is returned; the generated roxygen skeleton is
inserted into the document).
}
\description{
This function uses the OpenAI API to generate a roxygen skeleton for the
current selection in RStudio. The roxygen skeleton is then inserted into the
document using prompt: "insert roxygen skeleton to document this function"
}
