% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{list_index}
\alias{list_index}
\title{List Index Files}
\usage{
list_index()
}
\value{
A character vector containing the names of the index files found in
the specified directory.
}
\description{
This function lists the index files in the specified directory.
}
\examples{
\dontrun{
list_index()
}
}
