% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{save_user_config}
\alias{save_user_config}
\title{Save user configuration settings for gpttools}
\usage{
save_user_config(
  service = "openai",
  model = "gpt-4-turbo-preview",
  task = "Permissive Chat",
  local_embed = "Yes",
  openai_embed_model = "text-embedding-3-small",
  local_embed_model = "BAAI/bge-small-en-v1.5",
  k_context = 4,
  k_history = 4,
  save_history = "Yes",
  sources = "All",
  run_code = "No",
  persist = TRUE
)
}
\arguments{
\item{service}{The name of the service to use, default is "openai".}

\item{model}{The model to use, default is "gpt-4-1106-preview".}

\item{task}{The task to perform, default is "Permissive Chat".}

\item{local_embed}{Whether to use local embedding model. Default is "yes".}

\item{openai_embed_model}{The OpenAI embeddings model to use, default is
"text-embedding-3-small".}

\item{local_embed_model}{The local embeddings model to use, default is
"BAAI/bge-small-en-v1.5".}

\item{k_context}{The amount of context to keep, default is 4.}

\item{k_history}{The amount of history to keep, default is 4.}

\item{save_history}{Logical indicating whether history should be saved,
default is TRUE.}

\item{sources}{The sources to use, default is "All".}

\item{run_code}{Whether to execute generated code with \code{reprex::reprex()}}

\item{persist}{Logical indicating whether to persist the settings, default
is TRUE.}
}
\value{
Invisible NULL.
}
\description{
This function saves the user's configuration settings for gpttools service
to a YAML file.
}
