% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe.R
\name{transcribe_audio}
\alias{transcribe_audio}
\title{Transcribe audio and write to index}
\usage{
transcribe_audio(file_path, source, link = NULL, prompt = NULL)
}
\arguments{
\item{file_path}{Character string specifying the path to the audio file.}

\item{source}{Character string specifying the source of the audio.}

\item{link}{Character string specifying the link to the audio file
(optional).}

\item{prompt}{Character string specifying the prompt for transcription
(optional).}
}
\value{
The function writes an index in Parquet format to disk.
}
\description{
This function transcribes audio files, processes the text, and writes the
result to an index.
}
